+++ lib/directives/index.dart
@@ -4,6 +4,7 @@ export 'native_events.dart';
 export 'tabs.dart';
 export 'highlightjs.dart';
 export 'clipboard.dart';
+export 'showdownjs.dart';
 
 import 'poppable.dart';
 import 'collapsible.dart';
@@ -11,6 +12,7 @@ import 'native_events.dart';
 import 'tabs.dart';
 import 'highlightjs.dart';
 import 'clipboard.dart';
+import 'showdownjs.dart';
 
 const ngAdminDirectives = [
   PoppableDirective,
@@ -21,5 +23,6 @@ const ngAdminDirectives = [
   WTabViewDirective,
   WTabsDirective,
   HljsDirective,
-  WClipboardDirective
+  WClipboardDirective,
+  ShowdownDirective
 ];